% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_log_volume.R
\docType{data}
\name{data_log_volume}
\alias{data_log_volume}
\title{Intraday log volume dataset calculating from the S&P 500}
\format{
26 * 104 data.frame dataset containing 26 bins of 104 days observations of 1 stock.
}
\source{
tbd
}
\usage{
data(data_log_volume)
}
\description{
Intraday log volume dataset calculating from the S&P 500
}
\keyword{dataset}
