% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdx_volume.R
\docType{data}
\name{fdx_volume}
\alias{fdx_volume}
\title{15-min intraday volume of FDX}
\format{
An xts object.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(fdx_volume)
}
\description{
An xts object including 15-min trading volume of FDX from 2019-07-01 to 2019-12-31.
}
\keyword{dataset}
