% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDX_volume_xts.R
\docType{data}
\name{FDX_volume_xts}
\alias{FDX_volume_xts}
\title{15-min intraday volume of FDX}
\format{
An xts object.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(FDX_volume_xts)
}
\description{
An xts object including 15-min trading volume of FDX from 2019-07-01 to 2019-12-31.
}
\keyword{dataset}
