% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_unimodel.R
\name{forecast_unimodel}
\alias{forecast_unimodel}
\title{Forecast One-bin-ahead Financial Intraday Signal via a Univariate State-Space Model}
\usage{
forecast_unimodel(data, unimodel, out.sample)
}
\arguments{
\item{data}{A n_bin * n_day matrix or an xts object storing intraday signal.}

\item{unimodel}{A "\code{unimodel}" object from function \code{fit_unimodel}.}

\item{out.sample}{Number of days from the end of the dataset for out-of-sample forecast.}
}
\value{
A list containing the following elements:
        \item{\code{original.signal}}{A vector of original intraday signal;}
        \item{\code{forecast.signal}}{A vector of forecast intraday signal;}
        \item{\code{components}}{A list of the three forecast components:
             \itemize{ \item{\code{forecast.daily}}
                       \item{\code{forecast.seasonal}}
                       \item{\code{forecast.dynamic}}}}   
        \item{\code{error}}{A list of three error measures:
             \itemize{ \item{\code{mae}}
                       \item{\code{mape}}
                       \item{\code{rmse}}}}
}
\description{
The one-bin-ahead forecast is mathematically denoted by \eqn{\hat{y}_{\tau+1} = E[y_{\tau+1}|\{y_{j}\}_{j=1}^{\tau}]}{y*(\tau+1) = E[y(\tau + 1) | y(j), j = 1, ... , \tau]}.
             Given the dataset, you need to indicate how many days from the end of the dataset to keep for out-of-sample forecast.
             Three measures are used to evaluate the forecasting performance:
             \itemize{\item{Mean absolute error (MAE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M|\hat{y}_{\tau} - y_{\tau}|}{\sum (|y*(\tau) - y(\tau)|) / M} ;}
                      \item{Mean absolute percent error (MAPE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M\frac{|\hat{y}_{\tau} - y_{\tau}|}{y_{\tau}}}{\sum (|y*(\tau) - y(\tau)| / y(\tau)) / M} ;}
                      \item{Root mean square error (RMSE):
                            \eqn{\sqrt{\sum_{\tau=1}^M\frac{\left(\hat{y}_{\tau} - y_{\tau}\right)^2}{M}}}{[\sum ((y*(\tau) - y(\tau))^2 / M)]^0.5} ,}
             }
             where \eqn{M} is the total number of out-of-sample bins.
}
\examples{
\dontrun{

data(aapl_volume)

# Fit on the first 104 days
aapl_volume_fit <- aapl_volume[, 1:104]
unimodel_obj <- fit_unimodel(aapl_volume_fit)

# forecast on last 20 days
forecast_result <- forecast_unimodel(aapl_volume, unimodel_obj, out.sample = 20)
}

}
