% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_volume.R
\docType{data}
\name{GE_volume}
\alias{GE_volume}
\title{15-min intraday volume of GE from 2019-01-02 to 2019-06-28}
\format{
26 * 124 matrix dataset including 15-min trading volume of GE within 124 trading days from 2019-01-02 to 2019-06-28.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(GE_volume)
}
\description{
15-min intraday volume of GE from 2019-01-02 to 2019-06-28
}
\keyword{dataset}
