% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_plot}
\alias{generate_plot}
\title{Plot Analysis and Forecast Result}
\usage{
generate_plot(analysis_forecast_result)
}
\arguments{
\item{analysis_forecast_result}{Analysis/forecast result from \code{decompose_volume}.}
}
\value{
A list of \code{patchwork} objects:
       \itemize{
       \item{\code{log_components}: }{Plot of components of intraday signal in their log10 scale;}
       \item{\code{original_and_resultant}: }{Plot of the original and the resulant signals.}}
}
\description{
generate_plot the results of analysis and forecast.
}
\examples{
\dontrun{

data(aapl_volume)
aapl_volume_training <- aapl_volume[, 1:104]
aapl_volume_testing <- aapl_volume[, 105:124]

# Obtain smoothing and forecasting result
model_fit <- fit_volume(aapl_volume_training)
analysis_result <- decompose_volume(purpose = "analysis", model_fit, aapl_volume_training)
forecast_result <- decompose_volume(purpose = "forecast", model_fit, aapl_volume_testing)

# Plot components
generate_plot(analysis_result)
generate_plot(forecast_result)
}
}
