% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_decomposition}
\alias{plot_decomposition}
\title{Create plot from `uniModelfilter` result}
\usage{
plot_decomposition(data, filter.result)
}
\arguments{
\item{data}{The same data used in the \code{\link{uniModelfilter}}.}

\item{filter.result}{Decomposition result as obtained from the function \code{uniModelfilter}.}
}
\description{
After performing a filtering step with \code{\link{uniModelfilter}} 
this function creates four plots (daily average part, intraday dynamic part, seasonality and intraday volume) 
from the input data and decomposition result.
}
\examples{
library(intradayModel)
# load the data
data("data_log_volume")

# define the uniModel
modelSpec <- uniModelSpec(fit = TRUE)

# fit the model
modelSpec.fit <- uniModelFit(data_log_volume, modelSpec, maxit = 1000, abstol = 1e-4, log.switch = TRUE)

# filter
components <- uniModelFilter(data_log_volume, uniModel.fit)

# plot
plot_decomposition(data_log_volume, components)
}
\seealso{
\code{\link{uniModelfilter}}
}
\author{
Shengjie Xiu and Yifan Yu
}
