% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelFilter.R
\name{uniModelFilter}
\alias{uniModelFilter}
\title{Decompose Financial Intraday Signal via a Univariate State-Space Model}
\usage{
uniModelFilter(data, uniModel)
}
\arguments{
\item{data}{Matrix of intraday signal of size n_bin * n_day without any missing values.}

\item{uniModel}{Univariate model list object with all parameters fixed.}
}
\value{
A list containing the following elements:
       \item{\code{components}}{A list of the three components:
             \itemize{ \item{\code{daily}: Daily component;}
                       \item{\code{seasonal}: Seasonal component;}
                       \item{\code{dynamic}}: Intraday dynamic component.}}
       \item{\code{plot}}{A plot of the original signal and the three components.}
}
\description{
A model with all parameters fixed can be used to decompose the financial intraday signal into daily, seasonal, 
             and intraday dynamic components. The daily component and intraday dynamic component at time \eqn{\tau} are the smoothed state estimate 
             conditioned on all the data, and denoted by \eqn{\mathbb{E}[\mathbf{x}_{\tau}|\{y_{\tau}\}_{\tau=1}^{N}]}{E[ x(\tau) | y(\tau), \tau = 1, ... , N ]}, 
             where \eqn{N} is the total number of bins in the dataset. The seasonal component has the value of 
             \eqn{\boldsymbol{\phi}}{\phi}.
             
             This function will produce the above three components with a plot.
}
\examples{
# filter AAPL_volume
data(AAPL_volume)
model <- uniModelSpec(fit = TRUE)
model_fitted <- uniModelFit(AAPL_volume, model, acceleration = TRUE, 
                  maxit = 1000, abstol = 1e-4, log.switch = TRUE)
filter_result <- uniModelFilter(AAPL_volume, model_fitted)

}
\references{
Chen, R., Feng, Y., and Palomar, D. (2016). Forecasting intraday trading volume: A kalman filter approach. Available at SSRN 3101695.
}
\seealso{
\code{\link{uniModelSpec}}
}
