% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelForecast.R
\name{uniModelForecast}
\alias{uniModelForecast}
\title{Forecast One-bin-ahead Financial Intraday Signal via a Univariate State-Space Model}
\usage{
uniModelForecast(data, uniModel, out.sample)
}
\arguments{
\item{data}{A n_bin * n_day matrix or an xts object storing intraday signal.}

\item{uniModel}{A univariate model list object from function \code{uniModelFit}.}

\item{out.sample}{Number of days from the end of the dataset for out-of-sample forecast.}
}
\value{
A list containing the following elements:
        \item{\code{real.signal}}{A vector of real intraday signal;}
        \item{\code{forecast.signal}}{A vector of forecast intraday signal;}
        \item{\code{components}}{A list of the three forecast components:
             \itemize{ \item{\code{forecast.daily}}
                       \item{\code{forecast.seasonal}}
                       \item{\code{forecast.dynamic}}}}   
        \item{\code{error}}{A list of three error measures:
             \itemize{ \item{\code{mae}}
                       \item{\code{mape}}
                       \item{\code{rmse}}}}
}
\description{
The one-bin-ahead forecast is mathematically denoted by \eqn{\hat{y}_{\tau+1} = \mathbb{E}[y_{\tau+1}|\{y_{j}\}_{j=1}^{\tau}]}{y*(\tau+1) = E[y(\tau + 1) | y(j), j = 1, ... , \tau]}.
             Given the dataset, you need to indicate how many days from the end of the dataset to keep for out-of-sample forecast.
             Three measures are used to evaluate the forecasting performance:
             \itemize{\item{Mean absolute error (MAE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M\lvert\hat{y}_{\tau} - y_{\tau}\rvert}{\sum (|y*(\tau) - y(\tau)|) / M} ;}
                      \item{Mean absolute percent error (MAPE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M\frac{\lvert\hat{y}_{\tau} - y_{\tau}\rvert}{y_{\tau}},}{\sum (|y*(\tau) - y(\tau)| / y(\tau)) / M} ;}
                      \item{Root mean square error (RMSE):
                            \eqn{\sqrt{\sum_{\tau=1}^M\frac{\left(\hat{y}_{\tau} - y_{\tau}\right)^2}{M}}}{[\sum ((y*(\tau) - y(\tau))^2 / M)]^0.5} ,}
             }
             where \eqn{M} is the total number of out-of-sample bins.
}
\examples{
\dontrun{
 # One-bin-ahead prediction on the last 20 days of GE_volume
data(AAPL_volume)
AAPL_fit <- AAPL_volume[, 1:104]

model_fitted <- uniModelFit(AAPL_fit, control = list(acceleration = TRUE))
forecast_result <- uniModelForecast(AAPL_volume, model_fitted, out.sample = 20)
}

}
