% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelPred.R
\name{uniModelPred}
\alias{uniModelPred}
\title{Predict one-step ahead trading volume via kalman filter}
\usage{
uniModelPred(data, uniModel, out.sample)
}
\arguments{
\item{data}{n_bin * n_day log trading volume data matrix or data.frame with no NA.}

\item{uniModel}{uniModel object from function \code{uniModelSpec}.}

\item{out.sample}{Number of days before the last for out of sample prediction.}
}
\value{
A numeric vector containing the prediction result.
}
\description{
This function will return the modeled value of \eqn{y_t} conditioned on the previous data 
(the data during the time \eqn{0} to \eqn{t-1}).
}
\examples{
library(intradayModel)
}
\seealso{
\code{\link{uniModelSpec}}
}
\author{
Shengjie Xiu and Yifan Yu
}
