% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelSmooth.R
\name{uniModelSmooth}
\alias{uniModelSmooth}
\title{Smoothing Financial Intraday Signal via a Univariate State-Space Model}
\usage{
uniModelSmooth(data, uniModel)
}
\arguments{
\item{data}{A n_bin * n_day matrix or an xts object storing intraday signal.}

\item{uniModel}{A univariate model list object from function \code{uniModelFit}.}
}
\value{
A list containing the following elements:
       \item{\code{real.signal}}{A vector of real intraday signal;}
       \item{\code{smooth.signal}}{A vector of smoothed intraday signal;}
       \item{\code{components}}{A list of the three smoothed components:
             \itemize{ \item{\code{smooth.daily}}
                       \item{\code{smooth.seasonal}}
                       \item{\code{smooth.dynamic}}}}
}
\description{
A model with all parameters fixed can be used to decompose the financial intraday signal into daily, seasonal, 
             and intraday dynamic components. The daily component and intraday dynamic component at time \eqn{\tau} are the smoothed state estimate 
             conditioned on all the data, and denoted by \eqn{\mathbb{E}[\mathbf{x}_{\tau}|\{y_{\tau}\}_{\tau=1}^{N}]}{E[ x(\tau) | y(\tau), \tau = 1, ... , N ]}, 
             where \eqn{N} is the total number of bins in the dataset. The seasonal component has the value of 
             \eqn{\boldsymbol{\phi}}{\phi}.
}
\examples{
\dontrun{
data(AAPL_volume)
model_fitted <- uniModelFit(AAPL_volume, control = list(acceleration = TRUE))
smooth_result <- uniModelSmooth(AAPL_volume, model_fitted)
}

}
\references{
Chen, R., Feng, Y., and Palomar, D. (2016). Forecasting intraday trading volume: A kalman filter approach. Available at SSRN 3101695.
}
