% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_aapl.R
\docType{data}
\name{volume_aapl}
\alias{volume_aapl}
\title{15-min Intraday Volume of AAPL}
\format{
A 26 * 124 matrix.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(volume_aapl)
}
\description{
A 26 * 124 matrix including 15-min trading volume of AAPL from 2019-01-02 to 2019-06-28.
}
\keyword{dataset}
