% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_fdx.R
\docType{data}
\name{volume_fdx}
\alias{volume_fdx}
\title{15-min Intraday Volume of FDX}
\format{
An \code{xts} object.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(volume_fdx)
}
\description{
An \code{xts} object including 15-min trading volume of FDX from 2019-07-01 to 2019-12-31.
}
\keyword{dataset}
