% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAPL_volume.R
\docType{data}
\name{AAPL_volume}
\alias{AAPL_volume}
\title{15-min intraday volume of AAPL from 2019-01-02 to 2019-06-28}
\format{
26 * 124 matrix dataset including 15-min trading volume of AAPL within 124 trading days from 2019-01-02 to 2019-06-28.
}
\source{
\href{https://www.barchart.com}{barchart}
}
\usage{
data(AAPL_volume)
}
\description{
15-min intraday volume of AAPL from 2019-01-02 to 2019-06-28
}
\keyword{dataset}
