% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{autoplot}
\alias{autoplot}
\title{Plot Smoothed/Forecast Components}
\usage{
autoplot(smooth_forecast_result)
}
\arguments{
\item{smooth_forecast_result}{Smoothing/forecasting result from \code{smooth_unimodel} or \code{forecast_unimodel}.}
}
\value{
A \code{patchwork} object composed of 4 patches.
}
\description{
Plot the components of smoothing/forecasting result in one figure.
}
\examples{
\dontrun{

data(aapl_volume)
aapl_volume_training <- aapl_volume[, 1:104]

# Obtain smoothing and forecasting result
model_fit <- fit_volume(aapl_volume_training)
smooth_result <- smooth_unimodel(aapl_volume_training, model_fit)
forecast_result <- forecast_unimodel(aapl_volume, model_fit, out_sample = 20)

# Plot smoothed and forecast components
plot_components(smooth_result)
plot_components(forecast_result)
}
}
