% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_plots}
\alias{generate_plots}
\title{Plot Analysis and Forecast Result}
\usage{
generate_plots(analysis_forecast_result)
}
\arguments{
\item{analysis_forecast_result}{Analysis/forecast result from \code{decompose_volume()} or \code{forecast_volume()}.}
}
\value{
A list of \code{patchwork} objects:
       \itemize{
       \item{\code{components}: }{Plot of components of intraday volume;}
       \item{\code{log_components}: }{Plot of components of intraday volume in their log10 scale;}
       \item{\code{original_and_smooth} / \code{original_and_forecast}: }{Plot of the original and the smooth/forecast volume.}}
}
\description{
Generate plots of the analysis and forecast results.
}
\examples{
\dontrun{

data(aapl_volume)
aapl_volume_training <- aapl_volume[, 1:104]
aapl_volume_testing <- aapl_volume[, 105:124]

# Obtain analysis and forecast result
model_fit <- fit_volume(aapl_volume_training)
analysis_result <- decompose_volume(purpose = "analysis", model_fit, aapl_volume_training)
forecast_result <- forecast_volume(model_fit, aapl_volume_testing)

# Plot the analysis and forecast result
generate_plots(analysis_result)
generate_plots(forecast_result)
}
}
\author{
Shengjie Xiu, Yifan Yu and Daniel P. Palomar
}
