% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intradayModel-package.R
\docType{package}
\name{intradayModel-package}
\alias{intradayModel-package}
\title{intradayModel: Modeling and Forecasting Financial Intraday Signals}
\description{
`intradayModel` package uses state-of-the-art state-space models to facilitate the modeling and forecasting of 
financial intraday signals. This package currently offers a univariate model for intraday trading volume, 
with new features on intraday volatility and multivariate models in development. 
It is a valuable tool for anyone interested in exploring intraday, algorithmic, and high-frequency trading.
}
\section{Functions}{

\code{\link{uniModelSpec}},
\code{\link{uniModelFit}}, 
\code{\link{uniModelPred}}, 
\code{\link{uniModelFilter}}
}

\section{Data}{

\code{\link{AAPL_volume}},
\code{\link{GE_volume}}
}

\section{Help}{

For a quick help see the README file:
\href{https://github.com/convexfi/intradayModeling/blob/master/README.md}{GitHub-README}.
}

\author{
Shengjie Xiu, Yifan Yu and Daniel P. Palomar
}
