% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_components}
\alias{plot_components}
\title{Plot Components of Smoothing/Forecasting Result}
\usage{
plot_components(smooth_forecast_result)
}
\arguments{
\item{smooth_forecast_result}{Smoothing/forecasting result from function \code{uniModelSmooth} or \code{uniModelForecast}.}
}
\description{
Plot the components of smoothing/forecasting result in one figure.
}
\examples{
\dontrun{
 # Plot smoothed components of AAPL
data(AAPL_volume)
model_fitted <- uniModelFit(AAPL_volume, control = list(acceleration = TRUE))
smooth_result <- uniModelSmooth(AAPL_volume, model_fitted)
plot_components(smooth_result)
}
}
