% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_performance}
\alias{plot_performance}
\title{Plot Smoothing/Forecasting Performance}
\usage{
plot_performance(smooth_forecast_result)
}
\arguments{
\item{smooth_forecast_result}{Smoothing/forecasting result from function \code{uniModelSmooth} or \code{uniModelForecast}.}
}
\description{
Compares the real signal with the smoothed/forecast signal in one plot.
}
\examples{
\dontrun{
 # Plot forecast performance on AAPL
data(AAPL_volume)
AAPL_fit <- AAPL_volume[, 1:104]

model_fitted <- uniModelFit(AAPL_fit, control = list(acceleration = TRUE))
forecast_result <- uniModelForecast(AAPL_volume, model_fitted, out.sample = 20)
plot_performance(forecast_result)
}
}
