% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_performance}
\alias{plot_performance}
\title{Plot Smoothing/Forecasting Performance}
\usage{
plot_performance(smooth_forecast_result)
}
\arguments{
\item{smooth_forecast_result}{Smoothing/forecasting result from \code{smooth_unimodel} or \code{forecast_unimodel}.}
}
\value{
A \code{patchwork} object composed of 1 patch.
}
\description{
Compares the original signal with the smoothed/forecast signal in one plot.
}
\examples{
\dontrun{

data(aapl_volume)
aapl_volume_fit <- aapl_volume[, 1:104]

# Obtain smoothing and forecasting result
unimodel_obj <- fit_unimodel(aapl_volume_fit)
smooth_result <- smooth_unimodel(aapl_volume_fit, unimodel_obj)
forecast_result <- forecast_unimodel(aapl_volume, unimodel_obj, out.sample = 20)

# Plot smoothing and forecasting performance
plot_performance(smooth_result)
plot_performance(forecast_result)
}
}
