% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelFit.R
\name{uniModelFit}
\alias{uniModelFit}
\title{Fit a uniModel via Maximum-Likelihood Estimation}
\usage{
uniModelFit(
  data,
  uniModel,
  maxit = 3000,
  abstol = 1e-04,
  log.switch = TRUE,
  acceleration = FALSE
)
}
\arguments{
\item{data}{log volume data matrix of size n_bin * n_day with no NA for fitting.}

\item{uniModel}{uniModel object from function uniModelSpec}

\item{acceleration}{}
}
\description{
Fit a uniModel via Maximum-Likelihood Estimation
}
\references{
R. Chen, Y. Feng, and D. Palomar, “Forecasting intraday trading volume: a kalman filter approach,” Available at SSRN 3101695, 2016.
}
\author{
Shengjie Xiu and Yifan Yu
}
