% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelFit.R
\name{uniModelFit}
\alias{uniModelFit}
\title{Fit a uniModel via Expectation-Maximization algorithm}
\usage{
uniModelFit(
  data,
  uniModel,
  maxit = 3000,
  abstol = 1e-04,
  log.switch = TRUE,
  acceleration = FALSE
)
}
\arguments{
\item{data}{n_bin * n_day log trading volume data matrix or data.frame with no NA for fitting.}

\item{uniModel}{uniModel object from function \code{uniModelSpec}.}

\item{maxit}{Maximum number of iterations (default is \code{3000}).}

\item{abstol}{Absolute tolerance on the objective function to be used as the stopping criteria (default is \code{1e-4}).}

\item{log.switch}{Logical value indicating whether to record the history of convergence progress. 
If \code{TRUE}, a \code{uniModel} object with convergence log is returned (default is \code{TRUE}).}

\item{acceleration}{Logical value indicating whether to use accelerated Expectation-Maximization (EM) algorithm. 
If \code{TRUE}, accelerated EM algorithm is used (default is \code{FALSE}).}
}
\value{
A list containing the following elements (if the algorithm converges):
\item{\code{par}}{Values of fixed parameters.}
\item{\code{fit_request}}{List of logical values indicating whether the parameters are fixed or not. 
                          If the EM algorithm converges, all components of \code{fit_request} are \code{FALSE}.}
\item{\code{par_log}}{List of parameter values during the convergence.} 
If the algorithm does not converge, the return is the original input argument uniModel with unfixed parameters.
}
\description{
The main function for fitting state-space models. 
Two kinds of Expectation-Maximization (EM) algorithms are provided.
}
\examples{
library(intradayModel)
# load the data
data("data_log_volume")

# define the uniModel
modelSpec <- uniModelSpec(fit = TRUE)

# fit the model
modelSpec.fit <- uniModelFit(data_log_volume, modelSpec, maxit = 1000, abstol = 1e-4, log.switch = TRUE)
}
\references{
R. Chen, Y. Feng, and D. Palomar, “Forecasting intraday trading volume: a kalman filter approach,” Available at SSRN 3101695, 2016.
}
\seealso{
\code{\link{uniModelSpec}}
}
\author{
Shengjie Xiu and Yifan Yu
}
