% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelPred.R
\name{uniModelPred}
\alias{uniModelPred}
\title{Predict One-bin-ahead Financial Intraday Signal via a Univariate State-Space Model}
\usage{
uniModelPred(data, uniModel, out.sample)
}
\arguments{
\item{data}{Matrix of intraday signal of size n_bin * n_day without any missing values.}

\item{uniModel}{Univariate model list object with all parameters fixed.}

\item{out.sample}{Number of days from the end of the dataset for out-of-sample prediction.}
}
\value{
A list containing the following elements:
        \item{\code{signal_pred}}{One-bin-ahead prediction of intraday signal.}
        \item{\code{signal_real}}{Real out-of-sample intraday signal.}
        \item{\code{measure}}{MAE, MAPE, RMSE of out-of-sample prediction performance.}
        \item{\code{plot}}{Plot of the prediction and real values.}
}
\description{
The one-bin-ahead prediction is mathematically denoted by \eqn{\hat{y}_{\tau+1} = \mathbb{E}[y_{\tau+1}|\{y_{j}\}_{j=1}^{\tau}]}{y*(\tau+1) = E[y(\tau + 1) | y(j), j = 1, ... , \tau]}.
             Given the dataset, you need to indicate how many days from the end of the dataset to keep for out-of-sample prediction.
             Three measures are used to evaluate the prediction performance:
             \itemize{\item{Mean absolute error (MAE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M\lvert\hat{y}_{\tau} - y_{\tau}\rvert}{\sum (|y*(\tau) - y(\tau)|) / M} ;}
                      \item{Mean absolute percent error (MAPE):
                            \eqn{\frac{1}{M}\sum_{\tau=1}^M\frac{\lvert\hat{y}_{\tau} - y_{\tau}\rvert}{y_{\tau}},}{\sum (|y*(\tau) - y(\tau)| / y(\tau)) / M} ;}
                      \item{Root mean square error (RMSE):
                            \eqn{\sqrt{\sum_{\tau=1}^M\frac{\left(\hat{y}_{\tau} - y_{\tau}\right)^2}{M}}}{[\sum ((y*(\tau) - y(\tau))^2 / M)]^0.5} ,}
             }
             where \eqn{M} is the total number of out-of-sample bins.
}
\examples{
# One-bin-ahead prediction on the last 20 days of AAPL_volume
data("AAPL_volume")
data <- AAPL_volume
data_train <- AAPL_volume[, 1:104]

model <- uniModelSpec(fit = TRUE)
model_fitted <- uniModelFit(data_train, model, acceleration = TRUE)
predict_result <- uniModelPred(data, model_fitted, out.sample = 20)

}
\seealso{
\code{\link{uniModelSpec}}
}
