% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniModelSpec.R
\name{uniModelSpec}
\alias{uniModelSpec}
\title{Define a uniModel}
\usage{
uniModelSpec(fit = FALSE, init.pars = NULL, fixed.pars = NULL)
}
\arguments{
\item{fit}{Logical value indicating whether the model need to be fitted. (default is \code{FALSE}).}

\item{init.pars}{List of initial values of unfixed parameters}

\item{fixed.pars}{List of values of fixed parameters.}
}
\value{
A list containing the following elements:
}
\description{
Define the state space model for log trading volume.
}
\references{
R. Chen, Y. Feng, and D. Palomar, “Forecasting intraday trading volume: a kalman filter approach,” Available at SSRN 3101695, 2016.
}
\author{
Shengjie Xiu and Yifan Yu
}
